pyproject.toml
[project]
name = "myworkflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.30",
    "python-dotenv>=1.1.1",
]
